Remove AI agents tab and functionality from UI
- Remove entire /agents directory with all pages and components
- Remove agent API functions and types
- Remove AI agent feature flags from constants
- Clean up routing and navigation logic
- Move ChatMarkdown component to knowledge base directory to preserve functionality
- Knowledge base functionality remains identical for users
- Fix shared schemas in user creation modal

The agents tab is replaced by the knowledge base functionality.
