--- frontend/src/components/pages/agents/details/chat/chat-loading-indicator.tsx
@@ -1,7 +1,0 @@
-export const ChatLoadingIndicator = () => {
-  return (
-    <div className="flex justify-center items-center h-[calc(100vh-200px)]">
-      <div className="animate-pulse text-slate-500">Loading messages...</div>
-    </div>
-  );
-};
