--- frontend/src/components/pages/agents/hubspot-modal.scss
@@ -1,146 +1,0 @@
-.hs-input {
-    width: 100% !important; /* Override Hubspot */
-    height: 2.5rem;
-    font-size: 1rem;
-    -webkit-padding-start: 1rem;
-    padding-inline-start: 1rem;
-    -webkit-padding-end: 1rem;
-    padding-inline-end: 1rem;
-    border-radius: 0.375rem;
-    min-width: 0px;
-    outline: 2px solid transparent;
-    outline-offset: 2px;
-    position: relative;
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    -ms-appearance: none;
-    appearance: none;
-    transition-property: background-color, border-color, color, fill, stroke,
-      opacity, box-shadow, transform;
-    transition-duration: 200ms;
-    --input-font-size: 1rem;
-    --input-padding: 1rem;
-    --input-border-radius: 0.375rem;
-    --input-height: 2.5rem;
-    border: 1px solid;
-    border-color: inherit;
-    background: inherit;
-    &:focus-visible {
-      z-index: 1;
-      border-color: hsla(218, 75%, 61%, 0.8);
-      box-shadow: 0 0 0 1px hsla(218, 75%, 61%, 0.8);
-    }
-  }
-  
-  .hs-form fieldset label span {
-    text-align: start;
-    font-size: var(--chakra-fontSizes-sm);
-    -webkit-margin-end: var(--chakra-space-3);
-    margin-inline-end: var(--chakra-space-3);
-    font-weight: 600;
-    transition-property: var(--chakra-transition-property-common);
-    transition-duration: var(--chakra-transition-duration-normal);
-    opacity: 1;
-    margin-bottom: 0.25px;
-    margin-right: 0px;
-  }
-  
-  .field.hs-form-field {
-    margin-bottom: 16px; /* a hack since the empty fields from hubspot show display: none so we cannot use gap + flex */
-  }
-  
-  .no-list {
-    list-style-type: none; /* do not show error message with pseudo ::marker */
-  }
-  
-  .hs-form-required {
-    margin-left: 4px; /* separate '*' character for required fields from labels */
-  }
-  
-  fieldset {
-    width: 100%;
-  }
-  
-  .field {
-    display: block;
-    text-align: start;
-    font-size: var(--chakra-fontSizes-sm);
-    -webkit-margin-end: var(--chakra-space-3);
-    margin-inline-end: var(--chakra-space-3);
-    font-weight: 600;
-    transition-property: var(--chakra-transition-property-common);
-    transition-duration: var(--chakra-transition-duration-normal);
-    opacity: 1;
-    margin-bottom: 0.25px;
-    margin-right: 0px;
-  }
-  
-  .hs-form {
-    width: 100%;
-  }
-  
-  .submitted-message p {
-    text-align: left !important; // Needs override since Hubspot doesn't have any ID for the submitted message.
-    padding-bottom: var(--input-padding);
-  }
-  
-  .hs_error_rollup ul li label {
-    padding-bottom: 4px;
-    display: flex;
-    -webkit-align-items: center;
-    -webkit-box-align: center;
-    -ms-flex-align: center;
-    align-items: center;
-    color: #e53e3e;
-    margin-top: 0.5rem;
-    font-size: 0.85rem;
-    line-height: normal;
-    font-weight: 600;
-  }
-  
-  .hs-input.hs-fieldtype-textarea {
-    min-height: 120px;
-    width: 100%;
-    height: var(--input-height);
-    font-size: var(--input-font-size);
-    -webkit-padding-start: var(--input-padding);
-    padding-inline-start: var(--input-padding);
-    -webkit-padding-end: var(--input-padding);
-    padding-inline-end: var(--input-padding);
-    border-radius: var(--input-border-radius);
-    min-width: 0px;
-    outline: 2px solid transparent;
-    outline-offset: 2px;
-    position: relative;
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    -ms-appearance: none;
-    appearance: none;
-    transition-property: var(--chakra-transition-property-common);
-    transition-duration: var(--chakra-transition-duration-normal);
-    padding-top: var(--chakra-space-2);
-    padding-bottom: var(--chakra-space-2);
-    min-height: var(--chakra-sizes-20);
-    line-height: var(--chakra-lineHeights-short);
-    vertical-align: top;
-    --input-font-size: var(--chakra-fontSizes-md);
-    --input-padding: var(--chakra-space-4);
-    --input-border-radius: var(--chakra-radii-md);
-    --input-height: var(--chakra-sizes-10);
-    border: 1px solid;
-    border-color: inherit;
-    background: inherit;
-  
-    &:focus-visible {
-      z-index: 1;
-      border-color: #3182ce;
-      box-shadow: 0 0 0 1px #3182ce;
-    }
-  }
-  
-  .hs-field-desc {
-    font-size: var(--chakra-fontSizes-sm);
-    margin-bottom: var(--chakra-space-2);
-    margin-top: calc(var(--chakra-space-1) * -1);
-  }
-
