+++ frontend/src/components/pages/knowledgebase/KnowledgeBase.EditTabs.tsx
@@ -81,7 +81,7 @@ import { rpcnSecretManagerApi } from '../../../state/backendApi';
 import { getMessageFieldMetadata } from '../../../utils/protobuf-reflection';
 import { ProtoDisplayField, ProtoInputField, ProtoTextareaField } from '../../misc/ProtoFormField';
 import { SingleSelect } from '../../misc/Select';
-import { ChatMarkdown } from '../agents/details/chat/chat-markdown';
+import { ChatMarkdown } from './chat/chat-markdown';
 import { SecretsQuickAdd } from '../rp-connect/secrets/Secrets.QuickAdd';
 
 const { ToastContainer, toast } = createStandaloneToast();
