+++ frontend/src/components/pages/knowledgebase/chat/night-owl-theme.ts
