+++ frontend/src/components/pages/rp-connect/Pipelines.Create.tsx
@@ -17,7 +17,6 @@ import {
   Box,
   Button,
   Link as ChLink,
-  createStandaloneToast,
   Flex,
   FormField,
   Heading,
